+++ frontend/src/typings.d.ts
@@ -15,7 +15,6 @@ declare module 'numeral';
 declare module '@antv/data-set';
 declare module 'mockjs';
 declare module 'react-fittext';
-declare module 'bizcharts-plugin-slider';
 
 // preview.pro.ant.design only do not use in your production ;
 // preview.pro.ant.design Dedicated environment variable, please do not use it in your project.